+++ src/Recyclarr.Core/Config/Parsing/ConfigYamlDataObjects.cs
@@ -18,6 +18,7 @@ public record QualityScoreConfigYaml
 public record CustomFormatConfigYaml
 {
     public IReadOnlyCollection<string>? TrashIds { get; init; }
+    public int? Score { get; init; }
     public IReadOnlyCollection<QualityScoreConfigYaml>? AssignScoresTo { get; init; }
 }